+++ src/hotspot/share/gc/shenandoah/shenandoahForwardingTable.hpp
@@ -114,7 +114,7 @@ class ShenandoahForwardingTable {
 
   static uint64_t hash(HeapWord* original, void* table);
 
-  uint64_t index_of(HeapWord* original) const;
+  size_t index_of(HeapWord* original) const;
 
   template<class Entry>
   void enter_forwarding(BitMap& used, HeapWord* original, HeapWord* forwardee);
